Change how we upload to Sermon Audio
Split the creation of the Sermon Audio sermon record and the
mp3 file upload.  This lets us upload the mp3 file in the background
so we don't slow our upload process (because we don't have to wait
for the sermon file upload)

This also update the mp3 tags (and image!) before we upload the mp3
file to sermon audio.